drivers: remoteproc: Add C906L controller for Sophgo CV1800B SoC
Add initial support for the C906L remote processor found in the Sophgo
CV1800B SoC. The C906L is an asymmetric core typically used to run an
RTOS. This driver enables firmware loading and start/stop control of the
C906L processor via the remoteproc framework.

The C906L and the main application processor can communicate through
mailboxes. Support for mailbox-based functionality will be added in
a separate patch.

Signed-off-by: Junhui Liu <junhui.liu@pigmoral.tech>