riscv: dts: anlogic: add Milianke MLKPAI FS01 board
Add support for the Milianke MLKPAI FS01 board based on the Anlogic
DR1V90 SoC. Currently, the board can boot to a console via UART1, which
is connected to the onboard serial chip and routed to the Type-C
interface.

A "no4lvl" bootarg is added to force kernel to use 3-level paging,
because the DR1V90 hangs when attempting to set SATP to 4-level paging
mode during early boot, as detected in the set_satp_mode() function.

Signed-off-by: Junhui Liu <junhui.liu@pigmoral.tech>